Include admin section
It is not possible to edit anything yet, because the github repository
is not linked to Netlify.

The explicit link to config.yml on the admin page should not be
necessary, because it's using the default value. However, without this
setting Netlify CMS looks for config.yml at the wrong location
(/config.yml instead of /admin/config.yml) on now.sh (and locally with
`now dev`) for some reason.